♻️refactor: 동일 책임의 두 클래스를 하나의 클래스로 병합
책임 : 웹소켓 내 인증 관련 처리 유틸
