+++ OneCoin/Server/src/main/java/OneCoin/Server/config/auth/utils/UserUtilsForWebSocket.java
@@ -1,22 +1,47 @@
-package OneCoin.Server.chat.utils;
+package OneCoin.Server.config.auth.utils;
 
 import OneCoin.Server.config.auth.jwt.JwtTokenizer;
-import OneCoin.Server.config.auth.utils.CustomAuthorityUtils;
 import OneCoin.Server.user.entity.Role;
-import lombok.RequiredArgsConstructor;
+import OneCoin.Server.user.entity.User;
+import OneCoin.Server.user.service.UserService;
+import lombok.AllArgsConstructor;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.stereotype.Component;
 
+import java.security.Principal;
 import java.util.List;
 import java.util.Map;
 
 @Component
-@RequiredArgsConstructor
-public class AuthUtilForWebsocket {
+@AllArgsConstructor
+public class UserUtilsForWebSocket {
+    private UserService userService;
     private final JwtTokenizer jwtTokenizer;
     private final CustomAuthorityUtils customAuthorityUtils;
+
+    public User extractUser(Principal principal) {
+        try {
+            Map<String, Object> claims = extractClaims(principal);
+            Long userId = ((Integer) claims.get("id")).longValue();
+            User user = userService.findUser(userId);
+            return user;
+        } catch (Exception e) {
+            return null;
+        }
+    }
+
+    public Map<String, Object> extractClaims(Principal principal) {
+        try {
+            Authentication authentication = (Authentication) principal;
+            Map<String, Object> claims = (Map) authentication.getPrincipal();
+            return claims;
+        } catch (Exception e) {
+            return null;
+        }
+    }
+
     public Authentication authenticate(String accessToken) {
         Map<String, Object> claims = verifyJws(accessToken);
         return getAuthentication(claims);
